chore: reorganize file structure
