+++ README.md
@@ -5,19 +5,19 @@
 ### Deploy MyToken.sol
 
 ```bash
-npx ts-node ./scripts/TokenizedBallot/MyTokenDeploy.ts
+npx ts-node ./scripts/MyToken/deploy.ts
 ```
 
 ### Deploy TokenizedBallot.sol
 
 ```bash
-npx ts-node /TokenizedBallot/TokenizedBallotDeploy.ts \
+npx ts-node /TokenizedBallot/deploy.ts \
 	--address=0xfed.. --block=1 --proposals==Chocollate,Vanilla,Strawberry
 ```
 
 ### Vote with Vote.sol
 
 ```bash
-npx ts-node /TokenizedBallot/Vote.ts \
+npx ts-node /TokenizedBallot/vote.ts \
   --address=0xa1e.. --proposal=Chocollate --amount=100
 ```
