+++ scripts/MyToken/deploy.ts
@@ -2,7 +2,7 @@ import {
   abi,
   bytecode,
 } from "../../artifacts/contracts/MyToken.sol/MyToken.json";
-import { createClient } from "./lib";
+import { createClient } from "../lib";
 
 async function main() {
   console.log("Deploying...");
